update structure of of the active sounds list
